Witness generation fix
When we went through newly-populated values, as in

    for &(watch, _) in &buffer.target_values

`watch` was not necessarily a representative, because it came from a `GeneratedValues`, whose `set_target` doesn't know about representatives.

To avoid adding extra lookups, we can modify `extend` to return representative indices (which it already has to look up), and iterate through those instead of the original `GeneratedValues`.