Update size-optimized proof test (#368)
The 2^12 change made this outdated. We no longer need to shrink degree (since normal recursive proofs are 2^12), so we can simplify a bit. We just boost the rate, then do a size-optimized proof. (Without doing the rate boost first, the final proof would be over 2^12.)

Configured for 93 bits security for now, but the PoW settings are low so that'll be easy to increase.

~45kb with current settings.