Add `Environment` for reading/writing chain data
While generating a witness for System Zero, we need a way to access various chain data. For example, to execute the `BALANCE` instruction we need to access an arbitrary account's balance.

In our MVP, we plan to reuse witness generation code for normal transaction execution, even when we're not generating a proof. This is very suboptimal, but in the short term it will help keep our codebase small and nimble.

With this in mind, we anticipate that this `Environment` could have several implementations:
- A mock environment which stores everything in memory, just for testing.
- An environment backed by a real node's database, for when validators execute (not prove) transactions.
- An "light" environment which has only the data required to prove a particular transaction. This may be useful if we have prover nodes which don't run nodes, but just receive the bits of chain data they need from a "leader" who does run a node.

I also just moved the witness generation methods, as they're going to become much longer and more complex when integrated with `Environment` to actually read code from it etc.