Check that Merkle proofs have the proper length and leaf size
In our Merkle tree, there's no special field to distinguish leaves from branch nodes, so collisions between them may be possible. In particular, a prover could take the digest of a branch node, and prove inclusion a leaf with the same location and digest.[^1] A fixed-depth Merkle tree avoids such attacks, but we forgot to actually validate depth when verifying a Merkle proof.

Relatedly, we weren't checking that a leaf has the proper size. I'm not sure if an attacker can do anything useful with a wrongly sized leaf, particularly if the above issue is fixed, but it seems safest to just enforce that the entire hash tree structure is proper.

These changes are for the native verifier. Our recursive circuit already has the proper hash structure baked into it.

[^1]: This is only possible if a Merkle tree has a leaf size of 4, which I don't think occurs in our standard config.