Relax restriction to a single looked table
In our use of plonky2, we have a few instances where we need both
several looked and several looking tables.  Think of these lookups as
many-to-many multiplexers, or a kind of bus.  Looking tables can be
thought of as pushing values with an obligation to process them onto the
bus, looked tables can be seen as pulling values to discharge that
obligation.

My first attempt changed from `looked_table` to `looked_tables`, but
then I noticed that thanks to logup, we don't need this complication:
looked_tables are just looking_tables with negative multiplicities.

Thus we can get a simpler system by just removing `looked_table`
completely, and giving people tools to negate the multiplicities of
their lookup tables.

We also preserve the old `CrossTableLookup::new`, which applies the
negation to the passed `looked_table` automatically to keep the API as
compatible as possible.